+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchModule.kt
@@ -4,16 +4,19 @@ import com.f2prateek.rx.preferences2.Preference
 import com.f2prateek.rx.preferences2.RxSharedPreferences
 import dagger.Module
 import dagger.Provides
+import org.simple.clinic.main.TypedPreference
+import org.simple.clinic.main.TypedPreference.Type.OverdueSearchHistory
 import org.simple.clinic.remoteconfig.ConfigReader
 
 @Module
 object OverdueSearchModule {
 
   @Provides
-  fun overdueSearchHistory(
+  @TypedPreference(OverdueSearchHistory)
+  fun overdueSearchHistoryPreference(
       rxSharedPreferences: RxSharedPreferences
-  ): Preference<Set<String>> {
-    return rxSharedPreferences.getStringSet("preference_overdue_search_history_v1")
+  ): Preference<String> {
+    return rxSharedPreferences.getString("preference_overdue_search_history_v1")
   }
 
   @Provides